test(menu-toggle): migrate basic
